move go.mod and go.sum to backend
Moving `go.mod` and `go.sum` to the `backend` directory to keep all Go code in the same place. Also need to move around protos to work with this change. The latest version of `bufbuild/protocolbuffers/dart`, `v22.0.1` introduced a breaking change. Pinning to `v21.1.2` for now.
