+++ backend/api/proto/gen/go/tattooapp/v1/pbconnect/user_service.connect.go
@@ -8,7 +8,7 @@ import (
 	connect "connectrpc.com/connect"
 	context "context"
 	errors "errors"
-	v1 "github.com/kbremont/tattoo-app/api/proto/gen/go/tattooapp/v1"
+	v1 "github.com/kbremont/tattoo-app/backend/api/proto/gen/go/tattooapp/v1"
 	http "net/http"
 	strings "strings"
 )
